requirements.txt
langchain-core==0.3.8
langchain==0.3.2
langchain-community==0.3.1
langchain-google-genai==2.0.0
langchain-openai==0.2.1
pandas==2.2.2
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
streamlit==1.37.1
streamlit-option-menu==0.3.13
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
duckduckgo-search==6.3.0
streamlit_cookies_manager==0.2.0